extract worker types into separate classes (wip)
